fix: 리뷰 3건 — 저장 점수 읽기 계약 분리, 위젯 URI v6, reasoning.effort 문구 정정 (#19)
- 공유 FitAssessmentSchema.score는 0-100 실수 유지(이전 버전이 저장한 0.65 등 소수 값을
  그대로 읽음, 임의 변환 없음). 정수 제약은 생성 계약(AssessmentDraftSchema)에만 두고
  toAssessment가 전송 방식과 무관하게 재검사
- company 선택값은 v5 위젯의 strict 파서가 거부 → CAREER_RADAR_WIDGET_URI v6, README 안내
- reasoning.effort는 추론 강도 힌트이지 토큰 상한이 아니며 실제 감소는 미검증으로 문구 정정
- tests: 소수 점수 스냅샷 읽기·지원 저장·위젯 스키마 통과 (310 → 311)

+++ packages/shared/src/index.ts
@@ -63,8 +63,10 @@ export const GapSchema = z.object({
 
 export const FitAssessmentSchema = z.object({
   verdict: VerdictSchema, confidence: ConfidenceSchema,
-  // score, when present, is an integer on a 0-100 scale (100 = strongest evidence-backed fit); never a 0-1 fraction.
-  resumeContortion: ResumeContortionSchema, score: z.number().int().min(0).max(100).optional(),
+  // Read contract: any stored 0-100 value is accepted, because snapshots saved before milestone-4b2-v2
+  // can hold fractions such as 0.65 whose intended scale is unknown and must not be rewritten. The
+  // generation contract (server/src/ai/contracts.ts) requires an integer on the stated 0-100 scale.
+  resumeContortion: ResumeContortionSchema, score: z.number().min(0).max(100).optional(),
   strongestMatches: z.array(EvidenceMatchSchema), gaps: z.array(GapSchema),
   hardBlockers: z.array(GapSchema), interviewRisks: z.array(z.string().min(1)),
   recommendation: z.string().min(1), missingInformation: z.array(z.string().min(1)),
